package-changes-service/pyproject.toml
[tool.poetry]
name = "package-changes-service"
version = "1.0.0"
description = "OpenWrt Package Changes Service"
authors = ["ASU Team"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pyyaml = "^6.0"
watchdog = "^3.0.0"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
